Regression Analysis of the Same
This version takes the original Ship Station plotter, runs a regression, and extracts the beta-1 Coefficient (the slope), annexes that to the initial table, and comes out with two files:  one with a chart & table, and the other with the coefficients enumerated.  

I did this so that the top-ranking items could be evaluated by volume, or by how well they are picking-up in sales.